requirements.txt
splunk-sdk>=1.7.2
langchain>=0.1.0
langchain-community>=0.0.11
langchain-text-splitters>=0.0.1
langchain-ollama>=0.0.2
chromadb>=0.4.0
sentence-transformers>=2.2.2
ollama>=0.1.0
python-dotenv>=1.0.0
atlassian-python-api>=3.34.0
requests>=2.31.0
mitreattack-python>=2.0.14